+++ astropy/units/core.py
@@ -31,7 +31,8 @@ __all__ = [
     'NamedUnit', 'IrreducibleUnit', 'Unit', 'def_unit',
     'CompositeUnit', 'PrefixUnit', 'UnrecognizedUnit',
     'get_current_unit_registry', 'set_enabled_units',
-    'add_enabled_units', 'dimensionless_unscaled']
+    'add_enabled_units', 'set_enabled_equivalencies',
+    'add_enabled_equivalencies', 'dimensionless_unscaled']
 
 
 def _flatten_units_collection(items):
@@ -63,19 +64,70 @@ def _flatten_units_collection(items):
     return result
 
 
+def _normalize_equivalencies(equivalencies):
+    """
+    Normalizes equivalencies, ensuring each is a 4-tuple of the form::
+
+    (from_unit, to_unit, forward_func, backward_func)
+
+    Parameters
+    ----------
+    equivalencies : list of equivalency pairs
+
+    Raises
+    ------
+    ValueError if an equivalency cannot be interpreted
+    """
+    if equivalencies is None:
+        return []
+
+    normalized = []
+
+    for i, equiv in enumerate(equivalencies):
+        if len(equiv) == 2:
+            funit, tunit = equiv
+            a = b = lambda x: x
+        elif len(equiv) == 3:
+            funit, tunit, a = equiv
+            b = a
+        elif len(equiv) == 4:
+            funit, tunit, a, b = equiv
+        else:
+            raise ValueError(
+                "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
+        if not (isinstance(funit, UnitBase) and
+                (isinstance(tunit, UnitBase) or tunit is None) and
+                six.callable(a) and
+                six.callable(b)):
+            raise ValueError(
+                "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
+        normalized.append((funit, tunit, a, b))
+
+    return normalized
+
+
 class _UnitRegistry(object):
     """
     Manages a registry of the enabled units.
     """
-    def __init__(self, init=[]):
-        self._reset()
-        self.add_enabled_units(init)
+    def __init__(self, init=[], equivalencies=[]):
+        if isinstance(init, _UnitRegistry):
+            equivalencies = init.equivalencies
+            init = init.all_units
 
-    def _reset(self):
+        self._reset_units()
+        self._reset_equivalencies()
+        self.add_enabled_units(init)
+        self.add_enabled_equivalencies(equivalencies)
+
+    def _reset_units(self):
         self._all_units = set()
         self._non_prefix_units = set()
         self._registry = {}
         self._by_physical_type = {}
+
+    def _reset_equivalencies(self):
+        self._equivalencies = set()
 
     @property
     def registry(self):
@@ -105,7 +157,7 @@ class _UnitRegistry(object):
             methods like `UnitBase.find_equivalent_units` and
             `UnitBase.compose`.
         """
-        self._reset()
+        self._reset_units()
         return self.add_enabled_units(units)
 
     def add_enabled_units(self, units):
@@ -159,10 +211,54 @@ class _UnitRegistry(object):
             unit._get_physical_type_id(),
             set())
 
+    @property
+    def equivalencies(self):
+        return list(self._equivalencies)
+
+    def set_enabled_equivalencies(self, equivalencies):
+        """
+        Sets the equivalencies enabled in the unit registry.
+
+        These equivalencies are used if no explicit equivalencies are given,
+        both in unit conversion and in finding equivalent units.
+
+        This is meant in particular for allowing angles to be dimensionless.
+        Use with care.
+
+        Parameters
+        ----------
+        equivalencies : list of equivalent pairs
+            E.g., as returned by `astropy.units.angles_dimensionless`.
+        """
+        self._reset_equivalencies()
+        return self.add_enabled_equivalencies(equivalencies)
+
+    def add_enabled_equivalencies(self, equivalencies):
+        """
+        Adds to the set of equivalencies enabled in the unit registry.
+
+        These equivalencies are used if no explicit equivalencies are given,
+        both in unit conversion and in finding equivalent units.
+
+        This is meant in particular for allowing angles to be dimensionless.
+        Use with care.
+
+        Parameters
+        ----------
+        equivalencies : list of equivalent pairs
+            E.g., as returned by `astropy.units.angles_dimensionless`.
+        """
+        # pre-normalize list to help catch mistakes
+        equivalencies = _normalize_equivalencies(equivalencies)
+        self._equivalencies |= set(equivalencies)
+
 
 class _UnitContext(object):
-    def __init__(self, init=[]):
-        _unit_registries.append(_UnitRegistry(init=init))
+    def __init__(self, init=[], equivalencies=[]):
+        _unit_registries.append(
+            _UnitRegistry(
+                init=init,
+                equivalencies=equivalencies))
 
     def __enter__(self):
         pass
@@ -172,6 +268,8 @@ class _UnitContext(object):
 
 
 _unit_registries = [_UnitRegistry()]
+
+
 def get_current_unit_registry():
     return _unit_registries[-1]
 
@@ -218,7 +316,10 @@ def set_enabled_units(units):
       solRad       | 6.95508e+08 m   | R_sun, Rsun  ,
     ]
     """
-    context = _UnitContext()
+    # get a context with a new registry, using equivalencies of the current one
+    context = _UnitContext(
+        equivalencies=get_current_unit_registry().equivalencies)
+    # in this new current registry, enable the units requested
     get_current_unit_registry().set_enabled_units(units)
     return context
 
@@ -267,8 +368,66 @@ def add_enabled_units(units):
       yd           | 0.9144 m        | yard             ,
     ]
     """
-    context = _UnitContext(init=get_current_unit_registry().all_units)
+    # get a context with a new registry, which is a copy of the current one
+    context = _UnitContext(get_current_unit_registry())
+    # in this new current registry, enable the further units requested
     get_current_unit_registry().add_enabled_units(units)
+    return context
+
+
+def set_enabled_equivalencies(equivalencies):
+    """
+    Sets the equivalencies enabled in the unit registry.
+
+    These equivalencies are used if no explicit equivalencies are given,
+    both in unit conversion and in finding equivalent units.
+
+    This is meant in particular for allowing angles to be dimensionless.
+    Use with care.
+
+    Parameters
+    ----------
+    equivalencies : list of equivalent pairs
+        E.g., as returned by `astropy.units.angles_dimensionless`.
+
+    Examples
+    --------
+    Exponentiation normally requires dimensionless quantities.  To avoid
+    problems with complex phases::
+
+        >>> from astropy import units as u
+        >>> with u.set_enabled_equivalencies(u.dimensionless_angles()):
+        ...     phase = 0.5 * u.cycle
+        ...     np.exp(1j*phase)                          # doctest: +ELLIPSIS
+        <Quantity (-1+...j) >
+    """
+    # get a context with a new registry, using all units of the current one
+    context = _UnitContext(get_current_unit_registry().all_units)
+    # in this new current registry, enable the equivalencies requested
+    get_current_unit_registry().set_enabled_equivalencies(equivalencies)
+    return context
+
+
+def add_enabled_equivalencies(equivalencies):
+    """
+    Adds to the equivalencies enabled in the unit registry.
+
+    These equivalencies are used if no explicit equivalencies are given,
+    both in unit conversion and in finding equivalent units.
+
+    This is meant in particular for allowing angles to be dimensionless.
+    Since no equivalencies are enabled by default, generally it is recommended
+    to use `set_enabled_equivalencies`.
+
+    Parameters
+    ----------
+    equivalencies : list of equivalent pairs
+        E.g., as returned by `astropy.units.angles_dimensionless`.
+    """
+    # get a context with a new registry, which is a copy of the current one
+    context = _UnitContext(get_current_unit_registry())
+    # in this new current registry, enable the further equivalencies requested
+    get_current_unit_registry().add_enabled_equivalencies(equivalencies)
     return context
 
 
@@ -392,33 +551,27 @@ class UnitBase(object):
     @staticmethod
     def _normalize_equivalencies(equivalencies):
         """
-        Normalizes a list of equivalencies, by ensuring each element
-        is a 4-tuple of the form::
+        Normalizes equivalencies, ensuring each is a 4-tuple of the form::
 
-            (from_unit, to_unit, forward_func, backward_func)
+        (from_unit, to_unit, forward_func, backward_func)
 
-        Raises a ValueError if the equivalencies list is invalid.
+        Parameters
+        ----------
+        equivalencies : list of equivalency pairs, or `None`
+
+        Returns
+        -------
+        A normalized list, including possible global defaults set by, e.g.,
+        `set_enabled_equivalencies`, except when `equivalencies`=`None`,
+        in which case the returned list is always empty.
+
+        Raises
+        ------
+        ValueError if an equivalency cannot be interpreted
         """
-        normalized = []
-        for i, equiv in enumerate(equivalencies):
-            if len(equiv) == 2:
-                funit, tunit = equiv
-                a = b = lambda x: x
-            elif len(equiv) == 3:
-                funit, tunit, a = equiv
-                b = a
-            elif len(equiv) == 4:
-                funit, tunit, a, b = equiv
-            else:
-                raise ValueError(
-                    "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
-            if not (isinstance(funit, UnitBase) and
-                    (isinstance(tunit, UnitBase) or tunit is None) and
-                    six.callable(a) and
-                    six.callable(b)):
-                raise ValueError(
-                    "Invalid equivalence entry {0}: {1!r}".format(i, equiv))
-            normalized.append((funit, tunit, a, b))
+        normalized = _normalize_equivalencies(equivalencies)
+        if equivalencies is not None:
+            normalized += get_current_unit_registry().equivalencies
 
         return normalized
 
@@ -540,12 +693,16 @@ class UnitBase(object):
         Parameters
         ----------
         other : unit object or string or tuple
-           The unit to convert to. If a tuple of units is specified, this
-           method returns true if the unit matches any of the ones in the tuple.
+            The unit to convert to. If a tuple of units is specified, this
+            method returns true if the unit matches any of those in the tuple;
+            for this case, equivalencies are ignored.
 
         equivalencies : list of equivalence pairs, optional
-           A list of equivalence pairs to try if the units are not
-           directly convertible.  See :ref:`unit_equivalencies`.
+            A list of equivalence pairs to try if the units are not
+            directly convertible.  See :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         Returns
         -------
@@ -553,12 +710,21 @@ class UnitBase(object):
         """
 
         if isinstance(other, tuple):
-            return any(self.is_equivalent(u) for u in other)
+            return any(self.is_equivalent(u, equivalencies=None)
+                       for u in other)
         else:
             other = Unit(other, parse_strict='silent')
 
         equivalencies = self._normalize_equivalencies(equivalencies)
 
+        return self._is_equivalent(other, equivalencies)
+
+    def _is_equivalent(self, other, equivalencies=[]):
+        """Returns True if this unit is equivalent to `other`.
+        See `is_equivalent`, except that a proper Unit object should be
+        given (i.e., no string) and that the equivalency list should be
+        normalized using `_normalize_equivalencies`.
+        """
         if isinstance(other, UnrecognizedUnit):
             return False
 
@@ -578,8 +744,8 @@ class UnitBase(object):
                     except:
                         pass
                 else:
-                    if(unit.is_equivalent(a) and other.is_equivalent(b) or
-                       unit.is_equivalent(b) and other.is_equivalent(a)):
+                    if(unit._is_equivalent(a) and other._is_equivalent(b) or
+                       unit._is_equivalent(b) and other._is_equivalent(a)):
                         return True
 
         return False
@@ -608,11 +774,12 @@ class UnitBase(object):
                 except:
                     pass
             else:
-                if (unit.is_equivalent(funit) and other.is_equivalent(tunit)):
+                if unit._is_equivalent(funit) and other._is_equivalent(tunit):
                     scale1 = (unit / funit)._dimensionless_constant()
                     scale2 = (tunit / other)._dimensionless_constant()
                     return make_converter(scale1, a, scale2)
-                elif (other.is_equivalent(funit) and unit.is_equivalent(tunit)):
+                elif (other._is_equivalent(funit) and
+                      unit._is_equivalent(tunit)):
                     scale1 = (unit / tunit)._dimensionless_constant()
                     scale2 = (funit / other)._dimensionless_constant()
                     return make_converter(scale1, b, scale2)
@@ -642,11 +809,14 @@ class UnitBase(object):
         Parameters
         ----------
         other : unit object or string
-           The unit to convert to.
+            The unit to convert to.
 
         equivalencies : list of equivalence pairs, optional
-           A list of equivalence pairs to try if the units are not
-           directly convertible.  See :ref:`unit_equivalencies`.
+            A list of equivalence pairs to try if the units are not
+            directly convertible.  See :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         Returns
         -------
@@ -680,13 +850,16 @@ class UnitBase(object):
         other : unit object or string
             The unit to convert to.
 
-        value : scalar int or float, or sequence that can be converted to array, optional
+        value : scalar int or float, or sequence convertable to array, optional
             Value(s) in the current unit to be converted to the
             specified unit.  If not provided, defaults to 1.0
 
         equivalencies : list of equivalence pairs, optional
-           A list of equivalence pairs to try if the units are not
-           directly convertible.  See :ref:`unit_equivalencies`.
+            A list of equivalence pairs to try if the units are not
+            directly convertible.  See :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         Returns
         -------
@@ -765,9 +938,9 @@ class UnitBase(object):
         units = [unit]
         for funit, tunit, a, b in equivalencies:
             if tunit is not None:
-                if self.is_equivalent(funit):
+                if self._is_equivalent(funit):
                     units.append(tunit.decompose())
-                elif self.is_equivalent(tunit):
+                elif self._is_equivalent(tunit):
                     units.append(funit.decompose())
 
         # Store partial results
@@ -789,7 +962,7 @@ class UnitBase(object):
                 # without needing to do an exhaustive search.
                 if len(tunit_decomposed.bases) == 1:
                     for base, power in zip(u._bases, u._powers):
-                        if tunit_decomposed.is_equivalent(base):
+                        if tunit_decomposed._is_equivalent(base):
                             tunit = tunit ** power
                             tunit_decomposed = tunit_decomposed ** power
                             break
@@ -866,6 +1039,9 @@ class UnitBase(object):
         equivalencies : list of equivalence pairs, optional
             A list of equivalence pairs to also list.  See
             :ref:`unit_equivalencies`.
+            This list is in addition to possible global defaults set by, e.g.,
+            `set_enabled_equivalencies`.
+            Use `None` to turn off all equivalencies.
 
         units : set of units to compose to, optional
             If not provided, any known units may be used to compose
@@ -1095,6 +1271,8 @@ class UnitBase(object):
         equivalencies : list of equivalence pairs, optional
             A list of equivalence pairs to also list.  See
             :ref:`unit_equivalencies`.
+            Any list given, including an empty one, supercedes global defaults
+            that may be in effect (as set by `set_enabled_equivalencies`)
 
         units : set of units to search in, optional
             If not provided, all defined units will be searched for
@@ -1367,7 +1545,8 @@ class IrreducibleUnit(NamedUnit):
     def decompose(self, bases=set()):
         if len(bases) and not self in bases:
             for base in bases:
-                if self.is_equivalent(base):
+                # to avoid roundrip, ensure no default equivalencies get used
+                if self.is_equivalent(base, equivalencies=[]):
                     return CompositeUnit(self.to(base), [base], [1])
 
             raise UnitsError(
@@ -1410,8 +1589,8 @@ class UnrecognizedUnit(IrreducibleUnit):
             "with it are invalid.".format(self.name))
 
     __pow__ = __div__ = __rdiv__ = __truediv__ = __rtruediv__ = __mul__ = \
-               __rmul__ = __lt__ = __gt__ = __le__ = __ge__ = __neg__ = \
-               _unrecognized_operator
+        __rmul__ = __lt__ = __gt__ = __le__ = __ge__ = __neg__ = \
+        _unrecognized_operator
 
     def __eq__(self, other):
         other = Unit(other, parse_strict='silent')
@@ -1420,11 +1599,11 @@ class UnrecognizedUnit(IrreducibleUnit):
     def __ne__(self, other):
         return not (self == other)
 
-    def is_equivalent(self, other, equivalencies=[]):
+    def is_equivalent(self, other, equivalencies=None):
         self._normalize_equivalencies(equivalencies)
         return self == other
 
-    def get_converter(self, other, equivalencies=[]):
+    def get_converter(self, other, equivalencies=None):
         self._normalize_equivalencies(equivalencies)
         raise ValueError(
             "The unit {0!r} is unrecognized.  It can not be converted "
@@ -1494,15 +1673,15 @@ class _UnitMetaClass(type):
                 if parse_strict == 'silent':
                     pass
                 else:
-                    msg = "'{0}' did not parse as unit format '{1}': {2}".format(
-                        s, format, str(e))
+                    msg = ("'{0}' did not parse as unit format '{1}': {2}"
+                           .format(s, format, str(e)))
                     if parse_strict == 'raise':
                         raise ValueError(msg)
                     elif parse_strict == 'warn':
                         warnings.warn(msg, UnitsWarning)
                     else:
-                        raise ValueError(
-                            "'parse_strict' must be 'warn', 'raise' or 'silent'")
+                        raise ValueError("'parse_strict' must be 'warn', "
+                                         "'raise' or 'silent'")
                 return UnrecognizedUnit(s)
 
         elif isinstance(s, (int, float, np.floating, np.integer)):
@@ -1707,7 +1886,7 @@ class CompositeUnit(UnitBase):
         def add_unit(unit, power, scale):
             if unit not in bases:
                 for base in bases:
-                    if unit.is_equivalent(base):
+                    if unit._is_equivalent(base):
                         scale *= unit.to(base) ** power
                         unit = base
                         break